requirements.txt
coloredlogs==7.3
humanfriendly==4.4.1
importlib_metadata==6.7.0
isal==0.11.1
liftover==1.1.16
pydantic==2.4.2
pysam==0.16.0.1
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.5
typing_extensions==4.7.1
